requirements.txt
mysql-connector-python==8.0.5
Django==4.0.8